+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/index.ts
@@ -3,4 +3,5 @@ export * from './useCopyDocLink';
 export * from './useCreateChildDocTree';
 export * from './useDocUtils';
 export * from './useIsCollaborativeEditable';
+export * from './useMaskDocOption';
 export * from './useTrans';